+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/voice/InstructionDownloadTask.java
@@ -1,4 +1,4 @@
-package com.mapbox.services.android.navigation.ui.v5.voice.speech;
+package com.mapbox.services.android.navigation.ui.v5.voice;
 
 import android.os.AsyncTask;
